+++ examples/new-integration/src/index.ts
@@ -51,10 +51,8 @@ export class NewIntegration implements FullIntegration {
       })
     )
 
-    // TODO: check results for errors
-
     return {
-      status: { code: 0, details: { debugOutcome: 'success', results } },
+      status: { code: 0, details: [{ debugOutcome: 'success', results }] },
     }
   }